После выполнения настроек под 32bit алгоритм проверки цифровой подписи ГОСТ(NID_id_GostR3410_2001_CryptoPro_A_ParamSet) работает корректно

+++ src/main.cpp
@@ -58,7 +58,7 @@ uint64_t msec() {
 }
 
 void ms() {
-    printf("msec: %ld\n", msec() );
+    printf("msec: %u\n", (unsigned)msec() );
 }
 
 int main() {
